Add save and load mechanics
Export settings and savestates to a file on "save" action (Ctrl+S)
Load a savefile on "load" action (Ctrl+O).
At the moment, the current context is not saved.
A Savestate must be created prior to saving,
and must be loaded after loading the savefile

Created savestates are stored in a Map by default (can be prevented).
All savestates stored in this map are exported when creating a savefile.

Loading a savefile erases all previously-stored savestates.