Prompt skip on completed scenes
Prompt appears when reaching `goto *sXXX` with an already-read scene.
Prompt does not appear every time the `gameContext.label` is changed
to a scene label

Fix grahpic issue when skipping scenes:
script locks when graphic command does not incur change.

Fix issue of `if` commands not waiting for the sub-command to finish.