Prepare settings/save-states import/export
Update functions to export and import files
containing settings and save-states.

Exports and imports can contain multiple save-states,
and include the settings or not.

Both settings and save-states can be ignored on file import.

Multiple files can be imported at the same time.
The files are imported in their order of selection, new files
erasing common settings and save-states with older files.