Fix loading from title menu
The `Window` component now does not write context label and index
if it is already set. If it is not, it first tries to `quickLoad()`,
and if that fails, it starts from scene 20.

The "Start" button from title screen now sets the label and index
to the beginning of scene 20 before navigating to `Window`